# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API - Swagger Documentation",
        "contact": {
            "email": "simonsiva13@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8902/api/v1",
            "description": "Pet Shop API Server"
        }
    ],
    "paths": {
...
        "/product/create": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a product",
                "operationId": "create_product",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "schema": "CreateRequest",
                                "title": "Create Request",
                                "required": [
                                    "category_uuid",
                                    "brand_uuid",
                                    "title",
                                    "price",
                                    "description",
                                    "metadata"
                                ],
                                "properties": {
                                    "category_uuid": {
                                        "description": "Category UUID",
                                        "type": "string"
                                    },
                                    "brand_uuid": {
                                        "description": "Brand UUID",
                                        "type": "string"
                                    },
                                    "title": {
                                        "description": "Product title",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "number"
                                    },
                                    "description": {
                                        "description": "Product description",
                                        "type": "string"
                                    },
                                    "metadata": {
                                        "description": "Product metadata",
                                        "properties": {
                                            "image": {
                                                "type": "string"
                                            },
                                            "brand": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List all products",
                "operationId": "products",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "sortBy",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "description": "desc",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/product/{uuid}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Fetch a product",
                "operationId": "get_product",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "query",
                        "description": "uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Edit a product",
                "operationId": "edit_product",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "schema": "EditRequest",
                                "title": "Edit Request",
                                "required": [
                                    "category_uuid",
                                    "brand_uuid",
                                    "title",
                                    "price",
                                    "description",
                                    "metadata"
                                ],
                                "properties": {
                                    "category_uuid": {
                                        "description": "Category UUID",
                                        "type": "string"
                                    },
                                    "brand_uuid": {
                                        "description": "Brand UUID",
                                        "type": "string"
                                    },
                                    "title": {
                                        "description": "Product title",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "number"
                                    },
                                    "description": {
                                        "description": "Product description",
                                        "type": "string"
                                    },
                                    "metadata": {
                                        "description": "Product metadata",
                                        "properties": {
                                            "image": {
                                                "type": "string"
                                            },
                                            "brand": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "operationId": "delete_product",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "query",
                        "description": "uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}